requirements.txt
streamlit>=1.30.0
python-dotenv>=1.0.0
groq>=0.4.0
